requirements.txt
appdirs==1.4.3
asgiref==3.5.2
astroid==2.15.6
certifi==2019.11.28
chardet==3.0.4
dill==0.3.7
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==4.1
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
httpagentparser==1.9.1
ipaddr==2.2.0
isort==5.12.0
lazy-object-proxy==1.9.0
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
openpyxl==3.0.10
Pillow==9.2.0
platformdirs==3.9.1
progress==1.5
psycopg2==2.9.3
PyJWT==2.3.0
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyparsing==2.4.6
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
tzdata==2022.2
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
